+++ src/media_enrichment/__init__.py
@@ -1,7 +1,7 @@
 """media-enrichment Skill package.
 
-Version: 0.1.0-dev10
+Version: 0.1.0-dev11
 """
 
-__version__ = "0.1.0-dev10"
+__version__ = "0.1.0-dev11"
 SKILL_NAME = "media-enrichment"